chore: separate hooks
