+++ app/positions/components/PositionsContent.tsx
@@ -13,9 +13,9 @@ import EmptyScreen from '@/components/Status/EmptyScreen';
 import LoadingScreen from '@/components/Status/LoadingScreen';
 import { SupplyModal } from '@/components/supplyModal';
 import { WithdrawModal } from '@/components/withdrawModal';
-import useUserPositions from '@/hooks/useUserPositions';
 import { MarketPosition } from '@/utils/types';
 import { PositionsSummaryTable } from './PositionsSummaryTable';
+import useUserPositionsWithEarning from '@/hooks/useUserPositionsWithEarning';
 
 export default function Positions() {
   const [showSupplyModal, setShowSupplyModal] = useState<boolean>(false);
@@ -30,7 +30,7 @@ export default function Positions() {
     return account === address;
   }, [account, address]);
 
-  const { loading, isRefetching, data: marketPositions, refetch } = useUserPositions(account);
+  const { loading, isRefetching, data: marketPositions, refetch } = useUserPositionsWithEarning(account);
 
   const hasSuppliedMarkets = marketPositions && marketPositions.length > 0;
