+++ src/hooks/useUserPositions.ts
@@ -1,41 +1,10 @@
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
 
 import { useState, useEffect, useCallback } from 'react';
-import { Address } from 'viem';
 import { userPositionsQuery } from '@/graphql/queries';
 import { SupportedNetworks } from '@/utils/networks';
-import { MarketPosition, UserTransaction, UserTxTypes } from '@/utils/types';
+import { MarketPosition, UserTransaction } from '@/utils/types';
 import { getMarketWarningsWithDetail } from '@/utils/warnings';
-import { usePositionSnapshot } from './usePositionSnapshot';
-
-export type PositionEarnings = {
-  lifetimeEarned: string;
-  last24hEarned: string | null;
-  last7dEarned: string | null;
-  last30dEarned: string | null;
-};
-
-export function calculateEarningsFromSnapshot(
-  currentBalance: bigint,
-  snapshotBalance: bigint,
-  transactions: UserTransaction[],
-  start: number,
-  end: number = Math.floor(Date.now() / 1000),
-): string {
-  // Get transactions after snapshot timestamp
-  const txsWithinPeriod = transactions.filter((tx) => Number(tx.timestamp) > start && Number(tx.timestamp) < end);
-
-  const depositsAfter = txsWithinPeriod
-    .filter((tx) => tx.type === UserTxTypes.MarketSupply)
-    .reduce((sum, tx) => sum + BigInt(tx.data?.assets || '0'), 0n);
-
-  const withdrawsAfter = txsWithinPeriod
-    .filter((tx) => tx.type === UserTxTypes.MarketWithdraw)
-    .reduce((sum, tx) => sum + BigInt(tx.data?.assets || '0'), 0n);
-
-  const earned = currentBalance + withdrawsAfter - (snapshotBalance + depositsAfter);
-  return earned.toString();
-}
 
 const useUserPositions = (user: string | undefined, showEmpty = false) => {
   const [loading, setLoading] = useState(true);
@@ -43,64 +12,6 @@ const useUserPositions = (user: string | undefined, showEmpty = false) => {
   const [data, setData] = useState<MarketPosition[]>([]);
   const [history, setHistory] = useState<UserTransaction[]>([]);
   const [error, setError] = useState<unknown | null>(null);
-
-  const { fetchPositionSnapshot } = usePositionSnapshot();
-
-  const calculateEarningsFromPeriod = async (
-    position: MarketPosition,
-    transactions: UserTransaction[],
-    userAddress: Address,
-    chainId: number,
-  ) => {
-    const currentBalance = BigInt(position.supplyAssets);
-    const marketId = position.market.uniqueKey;
-
-    // Filter transactions for this specific market
-    const marketTxs = transactions.filter((tx) => tx.data?.market?.uniqueKey === marketId);
-
-    // Get historical snapshots
-    const now = Math.floor(Date.now() / 1000);
-    const snapshots = await Promise.all([
-      fetchPositionSnapshot(marketId, userAddress, chainId, now - 24 * 60 * 60), // 24h ago
-      fetchPositionSnapshot(marketId, userAddress, chainId, now - 7 * 24 * 60 * 60), // 7d ago
-      fetchPositionSnapshot(marketId, userAddress, chainId, now - 30 * 24 * 60 * 60), // 30d ago
-    ]);
-
-    const [snapshot24h, snapshot7d, snapshot30d] = snapshots;
-
-    return {
-      lifetimeEarned: calculateEarningsFromSnapshot(
-        currentBalance,
-        0n, // genesis snapshot: 0 balance
-        marketTxs,
-        0,
-      ),
-      last24hEarned: snapshot24h
-        ? calculateEarningsFromSnapshot(
-            currentBalance,
-            BigInt(snapshot24h.supplyAssets),
-            marketTxs,
-            now - 24 * 60 * 60,
-          )
-        : null,
-      last7dEarned: snapshot7d
-        ? calculateEarningsFromSnapshot(
-            currentBalance,
-            BigInt(snapshot7d.supplyAssets),
-            marketTxs,
-            now - 7 * 24 * 60 * 60,
-          )
-        : null,
-      last30dEarned: snapshot30d
-        ? calculateEarningsFromSnapshot(
-            currentBalance,
-            BigInt(snapshot30d.supplyAssets),
-            marketTxs,
-            now - 30 * 24 * 60 * 60,
-          )
-        : null,
-    };
-  };
 
   const fetchData = useCallback(
     async (isRefetch = false) => {
@@ -175,20 +86,12 @@ const useUserPositions = (user: string | undefined, showEmpty = false) => {
           marketPositions
             .filter((position: MarketPosition) => showEmpty || position.supplyShares.toString() !== '0')
             .map(async (position: MarketPosition) => {
-              const earnings = await calculateEarningsFromPeriod(
-                position,
-                transactions,
-                user as Address,
-                position.market.morphoBlue.chain.id,
-              );
-
               return {
                 ...position,
                 market: {
                   ...position.market,
                   warningsWithDetail: getMarketWarningsWithDetail(position.market),
                 },
-                earned: earnings,
               };
             }),
         );
@@ -203,7 +106,7 @@ const useUserPositions = (user: string | undefined, showEmpty = false) => {
         setIsRefetching(false);
       }
     },
-    [user, fetchPositionSnapshot, showEmpty],
+    [user, showEmpty],
   );
 
   useEffect(() => {
@@ -217,7 +120,7 @@ const useUserPositions = (user: string | undefined, showEmpty = false) => {
     [fetchData],
   );
 
-  return { loading, isRefetching, data, history, error, refetch, calculateEarningsFromPeriod };
+  return { loading, isRefetching, data, history, error, refetch };
 };
 
 export default useUserPositions;
